% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Btest.R
\name{Btest}
\alias{Btest}
\title{Bayes factors and posterior probabilities for linear regression models}
\usage{
Btest(
  models,
  data,
  prior.betas = "eZellnerSiow",
  g = dim(data)[1],
  prior.models = "Constant",
  priorprobs = NULL,
  null.model = NULL
)
}
\arguments{
\item{models}{A named list with the entertained models defined with their
corresponding formulas. If the list is unnamed, default names are given by
the routine. One model must be nested in all the others.}

\item{data}{data frame containing the data.}

\item{prior.betas}{Prior distribution for regression parameters within each
model (to be literally specified).}

\item{prior.models}{Type of prior probabilities of the models (to be literally specified). Possible choices are
"Constant" and "User" (see details).}

\item{priorprobs}{A named vector ir list (same length and names as in argument
\code{models}) with the prior probabilities of the models (used in combination
of \code{prior.models="User"}). If the provided object
is not named, then the order in the list of \code{models} is used to assign the prior
probabilities}

\item{null.model}{The name of the null model (eg. the one nested in all the others).
By default, the names of covariates in the different
models are used to identify the null model. An error is produced if such identification fails.
This identification
is not performed if the definition of the null model is provided, with this argument,
by the user.
Note that the (the \code{null.model} must coincide with that model with the largest
sum of squared errors and should be smaller in dimension to any other model).}
}
\value{
\code{Btest} returns an object of type \code{Btest} which is a
\code{list} with the following elements: \item{BFio }{A vector with the
Bayes factor of each model to the simplest model.} \item{PostProbi }{A
vector with the posterior probabilities of each model.} \item{models }{A
list with the entertained models. } \item{nullmodel}{The position of the
simplest model. }
\item{prior.betas}{prior.betas}
\item{g}{The value of g}
\item{prior.models}{prior.models}
\item{priorprobs}{priorprobs}
}
\description{
It Computes the Bayes factors and posterior probabilities of a list of linear
regression models proposed to explain a common response variable over the
same dataset
}
\details{
The Bayes factors, BFi0, are expressed in relation with the simplest model
(the one nested in all the others). Then, the posterior probabilities of the
entertained models are obtained as

Pr(Mi | \code{data})=Pr(Mi)*BFi0/C,

where Pr(Mi) is the prior probability of model Mi and C is the normalizing
constant.

The Bayes factor B_i depends on the prior assigned for the parameters in the regression

Option is "eZellnerSiow" a prior which 

With respect to the prior over the model space Pr(Mi) three possibilities
are implemented: "Constant", under which every model has the same prior
probability and "User". With this last option, the prior probabilities are
defined through the named list \code{priorprobs}. These probabilities can be
given unnormalized.
}
\references{
Bayarri, M.J., Berger, J.O., Forte, A. and Garcia-Donato, G.
(2012)<DOI:10.1214/12-aos1013> Criteria for Bayesian Model choice with
Application to Variable Selection. The Annals of Statistics. 40: 1550-1557.
}
\seealso{
\code{\link[BayesVarSel]{Bvs}} for variable selection within linear
regression models
}
\author{
Gonzalo Garcia-Donato and Anabel Forte

Maintainer: <anabel.forte@uv.es>
}
